pyproject.toml
[project]
name = "send-benchmarking"
version = "0.0.1"
description = "A SEND benchmarking web-tool using published data"
authors = [
    {name = "William Levack Payne",email = "108789355+WillLP-code@users.noreply.github.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "plotly (>=6.5.1,<7.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "django-crispy-forms (>=2.5,<3.0)",
    "django-select2 (>=8.4.8,<9.0.0)",
    "django-bootstrap5 (>=26.1,<27.0)",
    "django-bootstrap (>=0.2.4,<0.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
